Split out Mathics3 to SymPy predictate conversion
More tests.
